+++ previous-versions/files/codatfiles.go
@@ -71,7 +71,7 @@ func (c *sdkConfiguration) GetServerDetails() (string, map[string]string) {
 // [See our OpenAPI spec](https://github.com/codatio/oas)
 type CodatFiles struct {
 	// Endpoints to manage uploaded files.
-	Files *files
+	Files *Files
 
 	sdkConfiguration sdkConfiguration
 }
@@ -149,9 +149,9 @@ func New(opts ...SDKOption) *CodatFiles {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "3.0.0",
-			SDKVersion:        "0.25.0",
-			GenVersion:        "2.173.0",
-			UserAgent:         "speakeasy-sdk/go 0.25.0 2.173.0 3.0.0 github.com/codatio/client-sdk-go/previous-versions/files",
+			SDKVersion:        "0.26.0",
+			GenVersion:        "2.188.1",
+			UserAgent:         "speakeasy-sdk/go 0.26.0 2.188.1 3.0.0 github.com/codatio/client-sdk-go/previous-versions/files",
 		},
 	}
 	for _, opt := range opts {