+++ previous-versions/files/pkg/models/operations/downloadfiles.go
@@ -3,7 +3,6 @@
 package operations
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/files/pkg/models/shared"
 	"io"
 	"net/http"
 )
@@ -29,78 +28,16 @@ func (o *DownloadFilesRequest) GetDate() *string {
 	return o.Date
 }
 
-// DownloadFilesErrorMessage - You are using an outdated API key or a key not associated with that resource.
-type DownloadFilesErrorMessage struct {
-	// `True` if the error occurred transiently and can be retried.
-	CanBeRetried *string `json:"canBeRetried,omitempty"`
-	// Unique identifier used to propagate to all downstream services and determine the source of the error.
-	CorrelationID *string `json:"correlationId,omitempty"`
-	// Machine readable error code used to automate processes based on the code returned.
-	DetailedErrorCode *int64 `json:"detailedErrorCode,omitempty"`
-	// A brief description of the error.
-	Error *string `json:"error,omitempty"`
-	// Codat's service the returned the error.
-	Service *string `json:"service,omitempty"`
-	// The HTTP status code returned by the error.
-	StatusCode *int64 `json:"statusCode,omitempty"`
-}
-
-func (o *DownloadFilesErrorMessage) GetCanBeRetried() *string {
-	if o == nil {
-		return nil
-	}
-	return o.CanBeRetried
-}
-
-func (o *DownloadFilesErrorMessage) GetCorrelationID() *string {
-	if o == nil {
-		return nil
-	}
-	return o.CorrelationID
-}
-
-func (o *DownloadFilesErrorMessage) GetDetailedErrorCode() *int64 {
-	if o == nil {
-		return nil
-	}
-	return o.DetailedErrorCode
-}
-
-func (o *DownloadFilesErrorMessage) GetError() *string {
-	if o == nil {
-		return nil
-	}
-	return o.Error
-}
-
-func (o *DownloadFilesErrorMessage) GetService() *string {
-	if o == nil {
-		return nil
-	}
-	return o.Service
-}
-
-func (o *DownloadFilesErrorMessage) GetStatusCode() *int64 {
-	if o == nil {
-		return nil
-	}
-	return o.StatusCode
-}
-
 type DownloadFilesResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
 	// Success
 	// The Close method must be called on this field, even if it is not used, to prevent resource leaks.
 	Data io.ReadCloser
-	// You are using an outdated API key or a key not associated with that resource.
-	ErrorMessage *DownloadFilesErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
-	// The request made is not valid.
-	Schema *shared.Schema
 }
 
 func (o *DownloadFilesResponse) GetContentType() string {
@@ -117,13 +54,6 @@ func (o *DownloadFilesResponse) GetData() io.ReadCloser {
 	return o.Data
 }
 
-func (o *DownloadFilesResponse) GetErrorMessage() *DownloadFilesErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
-}
-
 func (o *DownloadFilesResponse) GetStatusCode() int {
 	if o == nil {
 		return 0
@@ -136,11 +66,4 @@ func (o *DownloadFilesResponse) GetRawResponse() *http.Response {
 		return nil
 	}
 	return o.RawResponse
-}
-
-func (o *DownloadFilesResponse) GetSchema() *shared.Schema {
-	if o == nil {
-		return nil
-	}
-	return o.Schema
 }